🐛 Fixed a bug
What Fixed:
🛠️ Bug in Move counts was fixed!
[DETAIL]
when the timer was starting by clicking on any of the timers then the move counts was working fine but when started through play button, move counter wasn't working the normal way.

Also,
On reset, the last active clock state was selected for getting started for the next round, has been fixed and starts the upper clock by default!